// src/interfaces/address.ts
export interface Address{
    cep: string;
    logradouro: string;
    complemento: string;
    bairro: string;
    unidade: string;
    localidade: string;
    uf: string;
    ibge: string;
    estado: string;
    regiao: string;
    gia: string;
    ddd: string;
    siafi: string;
    shippingConst:number;
}