// src/components/CEPForm/index.tsx
import { useState } from "react";
import { useCEPForm } from "./cep-form.schema";
import type { Address } from "../../interfaces/address"
import { formatCurrency } from "../../utils/formatCurrency";

const SHIPPING_BY_REGION: Record<string, number> = {
    "Norte": 39.90,
    "Nordeste": 29.90,
    "Centro-Oeste": 24.90,
    "Sudeste": 14.90,
    "Sul": 19.90,
}


export const CEPForm = () => {

    const {
        register,
        handleSubmit,
        errors,
        isSubmitting }
        = useCEPForm();

    const [address, setAddress] = useState<Address | null>(null);
    const [addressError, setAddressError] = useState<string | null>(null);



    const onSubmit = async ({ cep }: { cep: string }) => {

        setAddressError(null);
        setAddress(null);

        try {
            const response = await fetch(`https://viacep.com.br/ws/${cep}/json/`);
            const data = await response.json();

            if (data.erro) {
                setAddressError("CEP não encontrado");
                return;

            }
            const shippingConst = SHIPPING_BY_REGION[data.regiao];
            if (!shippingConst) {
                setAddressError("Região não encontrada");
                return;
            }

            setAddress({ ...data, shippingConst: shippingConst })
        } catch {
            setAddressError("Erro ao buscar CEP");
        }

    };
    return (
        <>
            <form className='flex gap-3' onSubmit={handleSubmit(onSubmit)} noValidate>
                <div className="flex flex-col flex-1">
                    <input type="text" placeholder="Insira seu CEP" className='border border-[#c0c0c0] rounded-md p-3 ${errors.cep ? "border-red-500" : ""}' {...register('cep')} />
                    {errors.cep && (
                        <span className='text-red-600 text-sm mt-1'>{String(errors.cep.message)}</span>
                    )}
                </div>
                <button type="submit" className='bg-black text-white py-3 px-6 rounded-md cursor-pointer hover:bg-gray-800 disabled:opacity-60' disabled={isSubmitting}>
                    Calcular
                </button>
            </form>
            {addressError &&(
                <div className="mt-4">
                    <p className="text-red-500 text-sm">{addressError}</p>
                </div>
            )}
            {address && (
                <div className="mt-4">
                    <p><strong>Região:</strong> {address.regiao}</p>
                    <p><strong>Custo de entrega:</strong> {formatCurrency(address.shippingConst)}</p>
                </div>
            )}
        </>
    );
};